requirements.txt
aiohttp==3.9.3
PyJWT==2.8.0
python-dotenv==1.0.1
cryptography==42.0.5
mcp[cli]>=0.1.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pytz==2025.2 